first iteration of functional frontend

+++ ATT-4008-4739/src/main/java/org/tardis/data/ASTCDP.java
@@ -13,7 +13,7 @@ public class ASTCDP extends DataPoint {
     private char[] ISO3;
     @Column(name = "Year", columnDefinition = "int")
     private int year;
-    @Column(name = "Value", columnDefinition = "int")
+    @Column(name = "Value", columnDefinition = "double")
     private double value;
